% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zip.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV Files}
\usage{
import_csv(csv_file, file_name, location)
}
\arguments{
\item{{zip_file}}{{A complete URL to the web-based csv file that ends in .csv}}

\item{{location}}{{Specify the directory you want to install into.}}
}
\description{
Take a URL for a CSV file and import into Posit Workbench. It will create the location folder if not specified,
otherwise will just save to the folder
}
