% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zip.R
\name{import_zip}
\alias{import_zip}
\title{Import Zip Files}
\usage{
import_zip(zip_file, name = "imported_data")
}
\arguments{
\item{{zip_file}}{{A complete URL to the web-based zip folder that ends in .zip}}

\item{{name}}{{Defaults to imported_data, but a string for the name of the folder being created}}
}
\description{
Take a URL for a zip folder and import into Posit Workbench
}
