% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_zip.R
\name{import_zip}
\alias{import_zip}
\title{Import Zip Files}
\usage{
import_zip(zip_file, location)
}
\arguments{
\item{{zip_file}}{{A complete URL to the web-based zip folder that ends in .zip}}

\item{{location}}{{Specify the directory you want to install into.}}
}
\description{
Take a URL for a zip folder and import into Posit Workbench. It will create the location folder if not specified,
otherwise will just save to the folder
}
